requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.2
django-ninja==1.4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==24.2
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
setuptools==75.8.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
wheel==0.45.1
